requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
av==17.0.1
certifi==2026.4.22
click==8.3.3
colorama==0.4.6
comtypes==1.4.16
ctranslate2==4.7.1
distro==1.9.0
docstring_parser==0.18.0
faster-whisper==1.2.1
filelock==3.29.0
flatbuffers==25.12.19
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.12.0
idna==3.13
jiter==0.14.0
keyboard==0.13.5
markdown-it-py==4.0.0
mdurl==0.1.2
MouseInfo==0.1.3
numpy==2.4.4
onnxruntime==1.25.1
packaging==26.2
protobuf==7.34.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.13.3
pydantic_core==2.46.3
PyGetWindow==0.0.9
Pygments==2.20.0
PyMsgBox==2.0.1
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.2.2
pyttsx3==2.99
pytweening==1.2.0
pywin32==311
PyYAML==6.0.3
rich==15.0.0
setuptools==82.0.1
shellingham==1.5.4
sniffio==1.3.1
tokenizers==0.23.1
tqdm==4.67.3
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
